Use ExprSplitter to eval incrementally in `includet` (fixes #475)
In `includet` mode you want to evaluate each statement.
However, when the same expression is repeated more than once,
the Dict key is identical and thus Revise keeps track of only one of
the expressions. This fixes the problem by evaluating every expression
as it is encountered.

This also significantly changes how expression-wrapping is done.
This new style should preserve more line numbers.