Allow methods to be defined in multiple files
This fixes cut/paste/save "user errors" in which a method is transiently
defined twice, once in the original file and once in a new location but
the user forgets to save the file after cutting it from the original
file. Revise will not insist on the file being saved, so if the user
never saves it there will be a dire warning when the package is next
precompiled, but that seems to be the least-bad outcome.

Formerly Revise would register the file as moved, and then when the
cut-file was saved it would delete the method.